+++ db/schema/users.sql
@@ -1,7 +1,7 @@
 CREATE TABLE users (
     user_id SERIAL NOT NULL PRIMARY KEY,
-    teleport_user_id BIGINT NOT NULL,
-    slack_user_id BIGINT NOT NULL,
+    teleport_user_id INT NOT NULL,
+    slack_user_id INT NOT NULL,
     FOREIGN KEY (teleport_user_id) REFERENCES teleport_users(teleport_user_id),
     FOREIGN KEY (slack_user_id) REFERENCES slack_users(slack_user_id),
     use_yn BOOLEAN NOT NULL DEFAULT TRUE,
